グローバルヘッダの追加
